requirements.txt
pandas
pyyaml
streamlit
plotly

